fio: fmap of files >4GB truncated the mapped size to uint32
builtin_map_file cast st.st_size (64-bit off_t) to uint32_t before handing it
to array_mark_locked — even though array_mark_locked takes a uint64 size and
Array::size is uint64. So fmap of any file >4GB set a wrapped size (e.g. an
8.5GB file -> 4.2GB), and reads past ~4.2GB threw 'array index out of range'.
Fixed by casting to uint64_t. Zero change for files <4GB (uint32==uint64 there).

Surfaced loading an 8.5GB Llama-3.1-8B GGUF in modules/dasLLAMA. (The fread-block
path at line ~446 truncates rlen to int32 separately — left as a known sibling.)

+++ src/builtin/module_builtin_fio.cpp
@@ -335,7 +335,9 @@ namespace das {
         fstat(fd, &st);
         void* data = mmap(nullptr, st.st_size, PROT_READ, MAP_SHARED, fd, 0);
         Array arr;
-        array_mark_locked(arr, data, uint32_t(st.st_size));
+        // st.st_size is 64-bit (off_t); array_mark_locked takes a uint64 size and
+        // Array::size is uint64 — a uint32 cast here truncated maps of files >4GB.
+        array_mark_locked(arr, data, uint64_t(st.st_size));
         vec4f args[1];
         args[0] = cast<Array *>::from(&arr);
         context->invoke(blk, args, nullptr, at);
